chore: add options to debug the nodeserver

+++ scripts/serverBuild.ts
@@ -5,8 +5,9 @@ import chokidar from "chokidar";
 import { log } from "./log";
 
 let childSpawn: any;
+type nodeArg = { argsBefore?: string[]; argsAfter?: string[] };
 
-function runNodeApp(launchJs: string) {
+function runNodeApp(launchJs: string, nodeArgs?: nodeArg) {
   function spawner(cmd: string, args: string[]) {
     childSpawn = spawn(cmd, args, {
       stdio: "inherit",
@@ -16,21 +17,39 @@ function runNodeApp(launchJs: string) {
       log(`\nNode app failed:${code}\n`);
     });
   }
-  spawner(process.platform === "win32" ? "node.exe" : "node", [launchJs]);
+
+  let args = [launchJs];
+  if (nodeArgs && nodeArgs.argsBefore) {
+    args = nodeArgs.argsBefore.concat(args);
+  }
+
+  if (nodeArgs && nodeArgs.argsBefore) {
+    args = args.concat(nodeArgs.argsBefore);
+  }
+
+  spawner(process.platform === "win32" ? "node.exe" : "node", args);
 }
 
-export async function server(watch: string, esbuildConfig: BuildOptions) {
+export async function server(
+  watch: string,
+  startNodejs: boolean,
+  esbuildConfig: BuildOptions,
+  nodeArgs?: nodeArg
+) {
   const builder = await build(esbuildConfig);
 
   chokidar.watch(watch, {}).on("change", async (eventName, path) => {
-    const msg =`client file changed ${eventName}`
+    const msg = `client file changed ${eventName}`;
     log(msg);
 
     // rebuild only be if incremental config
     if (builder.rebuild) {
       return builder.rebuild().then(() => {
-        childSpawn.kill();
-        if (esbuildConfig.outfile) {
+        if (childSpawn) {
+          childSpawn.kill();
+        }
+
+        if (esbuildConfig.outfile && startNodejs) {
           runNodeApp(esbuildConfig.outfile);
         }
 
@@ -42,7 +61,7 @@ export async function server(watch: string, esbuildConfig: BuildOptions) {
     }
   });
 
-  if (esbuildConfig.outfile) {
-    runNodeApp(esbuildConfig.outfile);
+  if (esbuildConfig.outfile && startNodejs) {
+    runNodeApp(esbuildConfig.outfile, nodeArgs);
   }
 }
